+++ sdk/communication/Azure.Communication.JobRouter/src/Models/CreateQueueOptions.cs
@@ -46,7 +46,7 @@ namespace Azure.Communication.JobRouter
         /// <summary>
         /// A set of key/value pairs that are identifying attributes used by the rules engines to make decisions.
         /// </summary>
-        public IDictionary<string, LabelValue> Labels { get; } = new Dictionary<string, LabelValue>();
+        public IDictionary<string, RouterValue> Labels { get; } = new Dictionary<string, RouterValue>();
 
         /// <summary>
         /// The content to send as the request conditions of the request.