+++ sdk/storage/Azure.Storage.DataMovement/src/StorageResourceItem.cs
@@ -184,17 +184,5 @@ namespace Azure.Storage.DataMovement
         /// Otherwise if the storage resource does not exist, false will be returned.
         /// </returns>
         protected internal abstract Task<bool> DeleteIfExistsAsync(CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Gets the source checkpoint data for this resource that will be written to the checkpointer.
-        /// </summary>
-        /// <returns>A <see cref="StorageResourceCheckpointData"/> containing the checkpoint information for this resource.</returns>
-        protected internal abstract StorageResourceCheckpointData GetSourceCheckpointData();
-
-        /// <summary>
-        /// Gets the destination checkpoint data for this resource that will be written to the checkpointer.
-        /// </summary>
-        /// <returns>A <see cref="StorageResourceCheckpointData"/> containing the checkpoint information for this resource.</returns>
-        protected internal abstract StorageResourceCheckpointData GetDestinationCheckpointData();
     }
 }