+++ sdk/storage/Azure.Storage.DataMovement/tests/Shared/MemoryStorageResourceItem.cs
@@ -68,14 +68,14 @@ namespace Azure.Storage.DataMovement.Tests
             throw new NotImplementedException();
         }
 
-        protected internal override StorageResourceCheckpointData GetDestinationCheckpointData()
-        {
-            throw new NotImplementedException();
-        }
-
         protected internal override Task<StorageResourceProperties> GetPropertiesAsync(CancellationToken token = default)
         {
             return Task.FromResult(new StorageResourceProperties(default, default, Buffer.Length, default));
+        }
+
+        protected internal override StorageResourceCheckpointData GetDestinationCheckpointData()
+        {
+            throw new NotImplementedException();
         }
 
         protected internal override StorageResourceCheckpointData GetSourceCheckpointData()